+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 v0.17.0 - _September 26, 2017_
+    * Added `zeroEx.exchange.validateOrderFillableOrThrowAsync` to simplify orderbook pruning (#170)
     * Made `zeroEx.exchange.getZRXTokenAddressAsync` public (#171)
 
 v0.16.0 - _September 20, 2017_